Conflicts:
	GameEngine/GameEngine/Engine/Threads/RenderScene.cpp

+++ GameEngine/GameEngine/Engine/Render/Key.hh
@@ -12,20 +12,20 @@ public:
 	Key();
 
 public:
-	static Key<TYPE> createKey(int index);
+	static Key<TYPE> createKey(size_t index);
 
 public:
-	int getId() const;
+	size_t getId() const;
 	operator bool() const;
 	void destroy();
 
 private:
-	int _id;
-	explicit Key(int id);
+	size_t _id;
+	explicit Key(size_t id);
 };
 
 template <typename type_t>
-int Key<type_t>::getId() const
+size_t Key<type_t>::getId() const
 {
 	return (_id);
 }
@@ -46,13 +46,13 @@ void Key<type_t>::destroy()
 }
 
 template <typename type_t>
-Key<type_t>::Key(int id)
+Key<type_t>::Key(size_t id)
 {
 	_id = id;
 }
 
 template <typename type_t>
-Key<type_t> Key<type_t>::createKey(int index)
+Key<type_t> Key<type_t>::createKey(size_t index)
 {
 	return (Key<type_t>(index));
 }
